+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Fixed
  - Fix Wine crash on newer Steam Runtime installations due to renamed runtime executable
+ - Fix graphical Wine applications crashing on Wayland
 
 ## [1.8.1] - 2022-03-20
 ### Added